Move pure Sandbox.Test cases into Sandbox.Test.Unit
