+++ engine/Sandbox.Test.Unit/MovieMaker/MovieTime.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Linq;
 using Sandbox.MovieMaker;
